# docs/openapi.json
{
  "components": {
    "schemas": {
      "AttachmentOut": {
        "properties": {
          "doc_id": {
            "default": "",
            "title": "Doc Id",
            "type": "string"
          },
          "file_name": {
            "default": "",
            "title": "File Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "upload_date": {
            "default": "",
            "title": "Upload Date",
            "type": "string"
          },
          "uploader_name": {
            "default": "",
            "title": "Uploader Name",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "AttachmentOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Browser-independent read-only facade for the authorized CTF target. Write operations are preview-only and never sent upstream. Local Bearer tokens never include upstream .iPSA cookie values.",
    "title": "iSStech Purchase Requisition Replay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/attachments/{attachment_id}/content": {
      "get": {
        "operationId": "download_attachment_v1_attachments__attachment_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "attachment_id",
            "required": true,
            "schema": {
              "title": "Attachment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "meta_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Meta Only",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Attachment",
        "tags": [
          "attachments"
        ]
      }
    },
...
    "/v1/purchase-requisitions/{requisition_id}/attachments": {
      "get": {
        "operationId": "list_attachments_v1_purchase_requisitions__requisition_id__attachments_get",
        "parameters": [
          {
            "in": "path",
            "name": "requisition_id",
            "required": true,
            "schema": {
              "title": "Requisition Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttachmentOut"
                  },
                  "title": "Response List Attachments V1 Purchase Requisitions  Requisition Id  Attachments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Attachments",
        "tags": [
          "attachments"
        ]
      }
    },
...
  }
}